Make it handle two bytes of button state data
Using two shift registers we are able to read two bytes of button states
and pass it via Serial to the python side. In python we split those 16
bits of button states into 4 groups of 4 buttons. So we have 4 callback
functions, that we can set, that can handle groups of four buttons.