Made CLK INH (Pin 15) of the 74HC165s LO
The input shift registers turned out to have a strong tendency to disturbance
and their signals were very unstable. The manual of 74HC165 says, that all
the unused input pins should be either at GND or VCC. For the CLK INH
(clock inhibit) VCC is not an option, as this would inhibit the clock.
Therefore we put them to GND.

It turns out to be far more stable.